+++ darq/app.py
@@ -1,3 +1,4 @@
+import asyncio
 import datetime
 import importlib
 import typing as t
@@ -11,8 +12,8 @@ from .jobs import Deserializer
 from .jobs import Job
 from .jobs import Serializer
 from .registry import Registry
-from .types import AnyCallable
 from .types import AnyTimedelta
+from .types import CoroutineType
 from .types import DataDict
 from .types import JobEnqueueOptions
 from .types import OnJobPostrunType
@@ -20,7 +21,7 @@ from .types import OnJobPrepublishType
 from .types import OnJobPrerunType
 from .utils import get_function_name
 from .utils import SecondsTimedelta
-from .worker import func as worker_func
+from .worker import Task
 from .worker import WorkerSettings
 
 CtxType = t.Dict[t.Any, t.Any]
@@ -43,35 +44,35 @@ class Darq:
     """
     Darq application.
 
-    :param queue_name: queue name to get jobs from
+    :param queue_name: queue name to get tasks from
     :param redis_settings: settings for creating a redis connection
-    :param burst: whether to stop the worker once all jobs have been run
+    :param burst: whether to stop the worker once all tasks have been run
     :param on_startup: coroutine function to run at worker startup
     :param on_shutdown: coroutine function to run at worker shutdown
-    :param on_job_prerun: coroutine function to run before job starts
-    :param on_job_postrun: coroutine function to run after job finish
-    :param on_job_prepublish: coroutine function to run before enqueue job
-    :param max_jobs: maximum number of jobs to run at a time
-    :param job_timeout: default job timeout (max run time)
-    :param keep_result: default duration to keep job results for
-    :param poll_delay: duration between polling the queue for new jobs
-    :param queue_read_limit: the maximum number of jobs to pull from the queue
+    :param on_job_prerun: coroutine function to run before task starts
+    :param on_job_postrun: coroutine function to run after task finish
+    :param on_job_prepublish: coroutine function to run before enqueue task
+    :param max_jobs: maximum number of tasks to run at a time
+    :param job_timeout: default task timeout (max run time)
+    :param keep_result: default duration to keep task results for
+    :param poll_delay: duration between polling the queue for new tasks
+    :param queue_read_limit: the maximum number of tasks to pull from the queue
                              each time it's polled;
                              by default it equals ``max_jobs``
-    :param max_tries: default maximum number of times to retry a job
+    :param max_tries: default maximum number of times to retry a task
     :param health_check_interval: how often to set the health check key
     :param health_check_key: redis key under which health check is set
     :param ctx: dict object, data from it will be pass to hooks:
                 ``on_startup``, ``on_shutdown`` - can modify ``ctx``;
                 ``on_job_prerun``, ``on_job_postrun`` - readonly
-    :param retry_jobs: whether to retry jobs on Retry or CancelledError or not
-    :param max_burst_jobs: the maximum number of jobs to process in burst mode
+    :param retry_jobs: whether to retry tasks on Retry or CancelledError or not
+    :param max_burst_jobs: the maximum number of tasks to process in burst mode
                            (disabled with negative values)
     :param job_serializer: a function that serializes Python objects to bytes,
                            defaults to pickle.dumps
     :param job_deserializer: a function that deserializes bytes into Python
                              objects, defaults to pickle.loads
-    :param default_job_expires: default job expires. If the job still hasn't
+    :param default_job_expires: default task expires. If the task still hasn't
                                 started after this duration, do not run it
     """
 
@@ -154,11 +155,11 @@ class Darq:
         :param cron_jobs: list of cron jobs to run,
                           use :func:`darq.cron.cron` to create them
         """
-        registered_tasks = {f.coroutine for f in self.registry.values()}
+        registered_coroutines = {f.coroutine for f in self.registry.values()}
         for cj in cron_jobs:
             if not isinstance(cj, CronJob):
                 raise DarqException(f'{cj!r} must be instance of CronJob')
-            if cj.task not in registered_tasks:
+            if cj.task not in registered_coroutines:
                 raise DarqException(
                     f'{cj.task!r} is not registered. '
                     'Please, wrap it with @task decorator.',
@@ -167,7 +168,7 @@ class Darq:
 
     def task(
             self,
-            func: t.Optional[AnyCallable] = None,
+            func: t.Optional[CoroutineType] = None,
             *,
             keep_result: t.Optional[AnyTimedelta] = None,
             timeout: t.Optional[AnyTimedelta] = None,
@@ -175,10 +176,24 @@ class Darq:
             queue: t.Optional[str] = None,
             expires: t.Optional[AnyTimedelta] = None,
     ) -> t.Any:
+        """
+        :param func: coroutine function
+        :param keep_result: duration to keep the result for, if 0 the result
+                            is not kept
+        :param timeout: maximum time the task should take
+        :param max_tries: maximum number of tries allowed for the task,
+                          use 1 to prevent retrying
+        :param queue: queue of the task, can be used to send task to
+                      different queue
+        :param expires: if the task still hasn't started after this
+                        duration, do not run it
+        """
         task_queue = queue
         task_expires = expires
 
-        def _decorate(function: AnyCallable) -> AnyCallable:
+        def _decorate(function: CoroutineType) -> CoroutineType:
+            assert asyncio.iscoroutinefunction(function), \
+                f'{function} is not a coroutine function'
             name = get_function_name(function)
 
             async def apply_async(
@@ -193,20 +208,20 @@ class Darq:
                     job_try: t.Optional[int] = None,
             ) -> t.Optional[Job]:
                 """
-                Enqueue a job with params.
+                Enqueue a task with params.
 
                 :param args: args to pass to the function
                 :param kwargs: any keyword arguments to pass to the function
-                :param job_id: ID of the job, can be used to enforce job
+                :param job_id: ID of the task, can be used to enforce task
                                uniqueness
-                :param queue: queue of the job, can be used to create job in
+                :param queue: queue of the task, can be used to send task to
                               different queue
-                :param defer_until: datetime at which to run the job
-                :param defer_by: duration to wait before running the job
-                :param expires: if the job still hasn't started after this
+                :param defer_until: datetime at which to run the task
+                :param defer_by: duration to wait before running the task
+                :param expires: if the task still hasn't started after this
                                 duration, do not run it
-                :param job_try: useful when re-enqueueing jobs within a job
-                :return: :class:`darq.jobs.Job` instance or ``None`` if a job
+                :param job_try: useful when re-enqueueing task
+                :return: :class:`darq.jobs.Job` instance or ``None`` if a task
                          with this ID already exists
                 """
                 args = list(args) if args is not None else []
@@ -242,11 +257,10 @@ class Darq:
 
             function.delay = delay  # type: ignore
             function.apply_async = apply_async  # type: ignore
-            arq_function = worker_func(
-                coroutine=function, name=name, default_queue=task_queue,
+            self.registry.add(Task.new(
+                coroutine=function, name=name,
                 keep_result=keep_result, timeout=timeout, max_tries=max_tries,
-            )
-            self.registry.add(arq_function)
+            ))
 
             return function